Extract a Contact class from Phone, Email and Address classes
